Feature/book rental crud

+++ src/main/java/com/example/Library/service/BookServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.Library.service;
 
+import com.example.Library.exception.BadRequestException;
 import com.example.Library.exception.NotFoundException;
 import com.example.Library.model.entity.Author;
 import com.example.Library.model.dto.BookDto;
@@ -45,7 +46,7 @@ public class BookServiceImpl implements BookService {
 
     public Book findBookModel(Long bookId){
         return bookRepository.findById(bookId)
-                .orElseThrow(() -> new NotFoundException("Book with id " + bookId + " is not found"));
+                .orElseThrow(() -> new NotFoundException(String.format("Book with id %s is not found", bookId)));
     }
 
     @Override
@@ -58,7 +59,14 @@ public class BookServiceImpl implements BookService {
     @Override
     public void delete(Long bookId) {
         Book book = findBookModel(bookId);
+        if(bookHasCopies(bookId)){
+            throw new BadRequestException(String.format("Book with id %s has bookCopies assigned to it", bookId));
+        }
         bookRepository.delete(book);
+    }
+
+    public boolean bookHasCopies(Long bookId) {
+        return bookRepository.bookHasCopies(bookId);
     }
 
     @Override
